Refactor YoutuberAdapter to its own source file
